docs/examples/cldk-demo/pyproject.toml
[tool.poetry]
name = "cldk-demo"
version = "0.2.0"
description = "Evaluation of Zircon models"
authors = ["Rahul Krishna <rkrsn@ibm.com>"]
readme = "README.md"


[tool.poetry.dependencies]
typer = "^0.12.3"
python = "^3.11"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
ibm-generative-ai = "^3.0.0"
yank-code = {git = "ssh://git@github.com/rahlk/yank-code.git", rev = "main"}
requests = "^2.32.3"
ollama = "^0.3.1"
cldk = {path = "/home/rkrsn/workspace/codellm-devkit", develop = true}

[tool.poetry.scripts]
java-testgen = "cldk_demo.java.test_generation:app"
java-summary = "cldk_demo.java.code_summary:app"
python-testgen = "cldk_demo.python.test_generation:app"
python-summary = "cldk_demo.python.code_summary:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
